New player. Ability to attack. Smoother idle, run, and jump transistions
Added a new player that has an attack animation. The player is a melee player. Updated the animations to have the player stop and run/walk instantly without a delay. Added animation parameter for the new player. Added an attack point so that the attack is only for a certain range